requirements.txt
amqp==2.3.2
billiard==3.5.0.3
cached-property==1.4.3
celery==4.2.0
certifi==2018.4.16
chardet==3.0.4
Django==2.0.5
djangorestframework==3.8.2
docker==3.4.1
docker-py==1.8.1
docker-pycreds==0.3.0
dockerpty==0.4.1
docopt==0.6.2
idna==2.6
jsonschema==2.6.0
kombu==4.2.1
Pygments==2.2.0
python-dateutil==2.4.2
pytz==2018.5
PyYAML==3.13
requests==2.18.4
six==1.11.0
structlog==15.3.0
texttable==0.9.1
urllib3==1.22
vine==1.1.4
websocket-client==0.48.0
psycopg2